requirements.txt
certifi==2018.11.29
chardet==3.0.4
defusedxml==0.5.0
diff-match-patch==20181111
Django==2.1.4
django-admin-rangefilter==0.3.12
django-import-export==1.2.0
et-xmlfile==1.0.1
gspread==3.1.0
httplib2==0.12.1
idna==2.8
jdcal==1.4
oauth2client==4.1.3
odfpy==1.4.0
openpyxl==2.6.0
pyasn1==0.4.5
pyasn1-modules==0.2.4
pytz==2018.7
PyYAML==3.13
requests==2.21.0
rsa==4.0
six==1.12.0
tablib==0.12.1
unicodecsv==0.14.1
urllib3==1.24.1
xlrd==1.2.0
xlwt==1.3.0
reportlab==3.5.34
django-filter==2.2.0
